backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
openai==1.35.13
google-generativeai==0.7.2
chromadb==0.5.5
numpy==1.26.4
PyMuPDF==1.24.9
python-docx==1.1.2
pypdf==4.3.1
aiohttp==3.10.5
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
black==24.8.0
flake8==7.1.1
httpx>=0.27.0